+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved README with license information and benchmark results
+- ErrGroup now throws `ObjectDisposedException` for any `Go(...)` call after disposal, preventing background failures when reusing disposed instances
 
 ## [1.0.0] - 2025-10-21